Reply with HttpClientConnection-Ref
The TCP.Connected instance isn't of much use to the client of the HttpClientConnection. However it would be good if a reference to the HttpClientConnection was available. The HostConnectorSetup returns such a self reference, albeit wrapped into a Http.HostConnectorInfo 

Also, it seems like a violation of the layering to return underlying Tcp object.